+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/di/FragmentModule.kt
@@ -3,7 +3,7 @@ package jp.chau2chaun2.honkot.sample.multimodule.di
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 import jp.chau2chaun2.honkot.sample.multimodule.feature.list_qiita_doc.ListQiitaDocFragment
-import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.SelectFragment
+import jp.chau2chaun2.honkot.sample.multimodule.feature.select.SelectFragment
 
 @Module
 abstract class FragmentModule {
