--- app/src/main/res/navigation/nav_main.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_main"
-    app:startDestination="@id/selectFragment">
-
-    <fragment
-        android:id="@+id/selectFragment"
-        android:name="jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.SelectFragment"
-        android:label="SelectFragment"
-        tools:layout="@layout/fragment_select">
-        <action
-            android:id="@+id/action_selectFragment_to_listDatabaseFragment"
-            app:destination="@id/listDatabaseFragment" />
-        <action
-            android:id="@+id/action_selectFragment_to_listAPIFragment"
-            app:destination="@id/listAPIFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/listAPIFragment"
-        android:name="jp.chau2chaun2.honkot.sample.multimodule.feature.list_qiita_doc.ListQiitaDocFragment"
-        android:label="ListAPIFragment"
-        tools:layout="@layout/fragment_list" />
-    <fragment
-        android:id="@+id/listDatabaseFragment"
-        android:name="jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragment"
-        android:label="ListDatabaseFragment"
-        tools:layout="@layout/fragment_list" />
-</navigation>
