+++ common/navigation/proguard-rules.pro
