+++ common/navigation/src/main/res/navigation/nav_main.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_main"
+    app:startDestination="@id/selectFragment">
+
+    <fragment
+        android:id="@+id/selectFragment"
+        android:name="jp.chau2chaun2.honkot.sample.multimodule.feature.select.SelectFragment"
+        android:label="SelectFragment">
+        <action
+            android:id="@+id/action_selectFragment_to_listDatabaseFragment"
+            app:destination="@id/listDatabaseFragment" />
+        <action
+            android:id="@+id/action_selectFragment_to_listAPIFragment"
+            app:destination="@id/listAPIFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/listAPIFragment"
+        android:name="jp.chau2chaun2.honkot.sample.multimodule.feature.list_qiita_doc.ListQiitaDocFragment"
+        android:label="ListAPIFragment" />
+    <fragment
+        android:id="@+id/listDatabaseFragment"
+        android:name="jp.chau2chaun2.honkot.sample.multimodule.feature.list_database.ListDatabaseFragment"
+        android:label="ListDatabaseFragment" />
+</navigation>
