+++ common/util/src/main/AndroidManifest.xml
