+++ common/util/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/util/IDataLoading.kt
