+++ feature/list_database/build.gradle
@@ -45,7 +45,7 @@ android {
 dependencies {
     implementation project(":data:model")
     implementation project(":data:repository")
-    implementation project(":util")
+    implementation project(":common:util")
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
