+++ settings.gradle
@@ -4,6 +4,9 @@ include ':data:api'
 include ':data:model'
 include ':data:dao'
 include ':data:repository'
-include ':util'
+include ':common:util'
+include ':common:navigation'
 include ':feature:list_database'
 include ':feature:list_qiita_doc'
+include ':feature:select'
+
